types: Moved Percentage from core to types
