+++ core/src/avm1/object/stage_object.rs
@@ -8,8 +8,8 @@ use crate::avm1::{Object, ObjectPtr, ScriptObject, TDisplayObject, TObject, Valu
 use crate::avm_warn;
 use crate::context::UpdateContext;
 use crate::display_object::{DisplayObject, EditText, MovieClip, TDisplayObjectContainer};
-use crate::types::Percent;
 use gc_arena::{Collect, GcCell, MutationContext};
+use ruffle_types::numbers::Percent;
 use ruffle_types::string::{AvmString, WStr};
 use std::fmt;
