+++ types/src/lib.rs
@@ -1,1 +1,2 @@
+pub mod numbers;
 pub mod string;
